make dragging cards actually affect game state
Rather than having one `<img>` for each card, there are now only enough cards
to fake it somewhat convincingly, and they just get reskinned as necessary.

Most of the tests are broken, but I'm leaving them as is for the moment, since
they're going to break some more soon with some restructuring.

Wars are not handled particularly well - that is, the first card played stays
on top in the playing field.

To prevent ninja-card-moving pre-empting our nice transitions, there are
`setTimeout` calls scattered liberally with constants.  We're blocking anyways,
so changing those to some form of `sleep()` would probably clean things up.